Cleaned up codebase, can download recusivley or in 1 directory!
Right now it has a GUI and only runs once. Next commit should be more bot oriented, and have it actively check what is on the board vs what is downloaded, and only download new files. Noticed a bug where it wouldn't mkdir would freak if it encountered a file with the same name. Have it run every x seconds for infinity depending on user input, and fix mkDir. Also have it only download new content after it has done its 1st run.

Then you can leave it run indefinitely and have an indefinite backup of the cesspool of the internet.